requirements.txt
Django
django-filter
django-tinymce
djangorestframework
djangorestframework-jwt
gunicorn
psycopg2-binary
